Complete all functions of lv_monitor
