+++ lvgl.h
@@ -62,7 +62,6 @@ extern "C" {
 #include "src/widgets/list/lv_list.h"
 #include "src/widgets/menu/lv_menu.h"
 #include "src/widgets/meter/lv_meter.h"
-#include "src/widgets/monitor/lv_monitor.h"
 #include "src/widgets/msgbox/lv_msgbox.h"
 #include "src/widgets/roller/lv_roller.h"
 #include "src/widgets/slider/lv_slider.h"
@@ -81,6 +80,7 @@ extern "C" {
 #include "src/others/gridnav/lv_gridnav.h"
 #include "src/others/fragment/lv_fragment.h"
 #include "src/others/imgfont/lv_imgfont.h"
+#include "src/others/monitor/lv_monitor.h"
 #include "src/others/msg/lv_msg.h"
 #include "src/others/ime/lv_ime_pinyin.h"
 #include "src/others/file_explorer/lv_file_explorer.h"
