+++ src/lv_conf_internal.h
@@ -849,7 +849,8 @@
  * Others
  *-----------*/
 
-/*1: Show CPU usage and FPS count*/
+/*1: Show CPU usage and FPS count
+ * Requires `LV_USE_MONITOR = 1*/
 #ifndef LV_USE_PERF_MONITOR
     #ifdef CONFIG_LV_USE_PERF_MONITOR
         #define LV_USE_PERF_MONITOR CONFIG_LV_USE_PERF_MONITOR
@@ -865,10 +866,18 @@
             #define LV_USE_PERF_MONITOR_POS LV_ALIGN_BOTTOM_RIGHT
         #endif
     #endif
+    #ifndef LV_USE_PERF_MONITOR_LOG_MDOE
+        #ifdef CONFIG_LV_USE_PERF_MONITOR_LOG_MDOE
+            #define LV_USE_PERF_MONITOR_LOG_MDOE CONFIG_LV_USE_PERF_MONITOR_LOG_MDOE
+        #else
+            #define LV_USE_PERF_MONITOR_LOG_MDOE 0
+        #endif
+    #endif
 #endif
 
 /*1: Show the used memory and the memory fragmentation
- * Requires `LV_USE_BUILTIN_MALLOC = 1`*/
+ * Requires `LV_USE_BUILTIN_MALLOC = 1`
+ * Requires `LV_USE_MONITOR = 1`*/
 #ifndef LV_USE_MEM_MONITOR
     #ifdef CONFIG_LV_USE_MEM_MONITOR
         #define LV_USE_MEM_MONITOR CONFIG_LV_USE_MEM_MONITOR
@@ -2316,6 +2325,19 @@
         #define LV_USE_SNAPSHOT CONFIG_LV_USE_SNAPSHOT
     #else
         #define LV_USE_SNAPSHOT 0
+    #endif
+#endif
+
+/*1: Enable Monitor component*/
+#ifndef LV_USE_MONITOR
+    #ifdef _LV_KCONFIG_PRESENT
+        #ifdef CONFIG_LV_USE_MONITOR
+            #define LV_USE_MONITOR CONFIG_LV_USE_MONITOR
+        #else
+            #define LV_USE_MONITOR 0
+        #endif
+    #else
+        #define LV_USE_MONITOR 1
     #endif
 #endif
